requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.0.1
click==8.1.7
email-validator==2.1.0
exceptiongroup==1.2.2
fastapi==0.112.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
passlib==1.7.4
pydantic==2.8.2
pydantic-core==2.20.1
PyJWT==2.9.0
python-multipart==0.0.6
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.2
typing-extensions==4.12.2
uvicorn==0.30.6
